refactoring HPO Association parser
