+++ phenol-annotations/src/main/java/org/monarchinitiative/phenol/annotations/assoc/HpoAssociationParser.java
@@ -51,7 +51,7 @@ public class HpoAssociationParser {
   private final File phenotypeDotHpoaFile;
   /** Key--an EntrezGene id; value--the corresponding symbol. all */
   private BiMap<TermId,String> allGeneIdToSymbolMap;
-  private ImmutableMap<TermId, String> geneIdToSymbolMap;
+  private Map<TermId, String> geneIdToSymbolMap;
   /** Key: an OMIM curie (e.g., OMIM:600100); value--corresponding GeneToAssociation object). For instance,
    * MICROVASCULAR COMPLICATIONS OF DIABETES, SUSCEPTIBILITY TO, 1; is associated to the gene VEGF as POLYGENIC,
    * and MARFAN SYNDROME is associated to the gene FBN1 as MENDELIAN.*/
@@ -83,7 +83,7 @@ public class HpoAssociationParser {
     this.mim2geneMedgenFile = new File(mim2geneMedgenPath);
     this.orphaToGeneFile = new File(orphaToGenePath);
     this.phenotypeDotHpoaFile = new File(phenotypeHpoaPath);
-    checkCoreFiles();
+
     checkOrphaFile();
     parse();
     ingestPhenotypeHpoaFile();
@@ -99,13 +99,13 @@ public class HpoAssociationParser {
     this.mim2geneMedgenFile = mim2geneMedgenFile;
     this.orphaToGeneFile = orphaToGeneFile;
     this.phenotypeDotHpoaFile = phenotypeHpoaFile;
-    checkCoreFiles();
+
     checkOrphaFile();
     parse();
     ingestPhenotypeHpoaFile();
   }
 
-  /** Parse everything except the Orphanet data!.*/
+  /** Parse everything except the Orphanet data!.
   public HpoAssociationParser(String geneInfoPath, String mim2geneMedgenPath, Ontology hpoOntology){
     this.hpoOntology = hpoOntology;
     this.homoSapiensGeneInfoFile = new File(geneInfoPath);
@@ -113,16 +113,8 @@ public class HpoAssociationParser {
     this.orphaToGeneFile = null;
     this.phenotypeDotHpoaFile = null;
     parse();
-  }
+  }*/
 
-  private void checkCoreFiles() {
-    if (! homoSapiensGeneInfoFile.exists()) {
-      throw new PhenolRuntimeException("Cannot find Homo_sapiens.gene_info.gz file");
-    }
-    if (! mim2geneMedgenFile.exists()) {
-      throw new PhenolRuntimeException("Cannot find mim2gene_medgen file");
-    }
-  }
 
   private void checkOrphaFile() {
     if (! orphaToGeneFile.exists()) {
@@ -242,91 +234,21 @@ public class HpoAssociationParser {
 
 
   private void parse() {
-
+    Gene2DiseaseAssociationParser parser =
+      new Gene2DiseaseAssociationParser(this.homoSapiensGeneInfoFile, this.mim2geneMedgenFile, this.orphaToGeneFile);
+    Multimap<TermId, GeneToAssociation> associationMap = parser.getAssociationMap();
+    //this.associationList = parser.getAssociationMap();
     ImmutableList.Builder<DiseaseToGeneAssociation> builder = new ImmutableList.Builder<>();
-
-    try {
-      this.allGeneIdToSymbolMap = parseGeneInfo(this.homoSapiensGeneInfoFile);
-      parseDiseaseToGene();
-      for (TermId omimCurie : associationMap.keySet()) {
-        Collection<GeneToAssociation> g2aList = associationMap.get(omimCurie);
-        DiseaseToGeneAssociation g2p = new DiseaseToGeneAssociation(omimCurie, ImmutableList.copyOf(g2aList));
-        builder.add(g2p);
-      }
-
-    } catch (IOException e) {
-      e.printStackTrace();
+    for (TermId omimCurie : associationMap.keySet()) {
+      Collection<GeneToAssociation> g2aList = associationMap.get(omimCurie);
+      DiseaseToGeneAssociation g2p = new DiseaseToGeneAssociation(omimCurie, ImmutableList.copyOf(g2aList));
+      builder.add(g2p);
     }
     this.associationList = builder.build();
+    this.geneIdToSymbolMap = parser.getGeneIdToSymbolMap();
     this.setAssociationMaps();
   }
 
-
-  public static Multimap<TermId, GeneToAssociation>  parseMim2GeneMedgen(BiMap<TermId,String> allGeneIdToSymbolMap, File mim2geneMedgenFile) throws IOException {
-    Multimap<TermId, GeneToAssociation> associationMap = ArrayListMultimap.create();
-    Map<TermId, String> geneMap = new HashMap<>();
-    try (BufferedReader br = new BufferedReader(new FileReader(mim2geneMedgenFile))) {
-      String line;
-      while ((line = br.readLine()) != null) {
-        if (line.startsWith("#")) continue;
-        String[] associations = line.split("\t");
-        if (associations[2].equals("phenotype")) {
-          String mimid = associations[0];
-          TermId omimCurie = TermId.of(OMIM_PREFIX, mimid);
-          String entrezGeneNumber = associations[1];
-          TermId entrezId = TermId.of(ENTREZ_GENE_PREFIX, entrezGeneNumber);
-          String symbol = allGeneIdToSymbolMap.get(entrezId);
-          if (!"-".equals(entrezGeneNumber)) {
-            if (symbol == null) {
-              symbol = "-";
-            } else {
-              if (!geneMap.containsKey(entrezId)) {
-                geneMap.put(entrezId, symbol);
-              }
-            }
-            TermId geneId = TermId.of(ENTREZ_GENE_PREFIX, entrezGeneNumber);
-            Gene gene = new Gene(geneId, symbol);
-            if (associations[5].contains("susceptibility")) {
-              GeneToAssociation g2a = new GeneToAssociation(gene, AssociationType.POLYGENIC);
-              if (!associationMap.containsEntry(omimCurie, g2a)) {
-                associationMap.put(omimCurie, g2a);
-              }
-            } else {
-              GeneToAssociation g2a = new GeneToAssociation(gene, AssociationType.MENDELIAN);
-              if (!associationMap.containsEntry(omimCurie, g2a)) {
-                associationMap.put(omimCurie, g2a);
-              }
-            }
-          }
-        }
-      }
-    }
-    return associationMap;
-  }
-
-  private void parseOrphaToGene() {
-    Multimap<TermId,String> orphaToGene;
-    if(this.orphaToGeneFile != null){
-      Map<String, TermId> geneSymbolToId = this.allGeneIdToSymbolMap.inverse();
-      try{
-        OrphaGeneToDiseaseParser parser = new OrphaGeneToDiseaseParser(this.orphaToGeneFile);
-        orphaToGene = parser.getOrphaDiseaseToGeneSymbolMap();
-        for (Map.Entry<TermId, String> entry : orphaToGene.entries()) {
-          TermId orpha = entry.getKey();
-          String geneSymbol = entry.getValue();
-          if(geneSymbolToId.containsKey(geneSymbol)){
-            Gene gene = new Gene(geneSymbolToId.get(geneSymbol), geneSymbol);
-            GeneToAssociation g2a = new GeneToAssociation(gene, AssociationType.UNKNOWN);
-            if(!associationMap.containsEntry(orpha,g2a)){
-              associationMap.put(orpha, g2a);
-            }
-          }
-        }
-      }catch(PhenolException e){
-        System.err.println(e.toString());
-      }
-    }
-  }
 
 
 
@@ -409,39 +331,6 @@ public class HpoAssociationParser {
     this.allGeneIdToSymbolMap = null;
   }
 
-  /**
-   * Parse the NCBI Homo_sapiens_gene_info.gz file
-   * Add the mappings to a Guava bimap, e.g., NCBIGene:150-ADRA2A
-   * @throws IOException if the file cannot be read
-   */
-  private BiMap<TermId,String> parseGeneInfo(File homoSapiensGeneInfoFile) throws IOException {
-    ImmutableBiMap.Builder<TermId,String> builder=new ImmutableBiMap.Builder<>();
-    InputStream fileStream = new FileInputStream(homoSapiensGeneInfoFile);
-    InputStream gzipStream = new GZIPInputStream(fileStream);
-    Reader decoder = new InputStreamReader(gzipStream);
-    BufferedReader br = new BufferedReader(decoder);
-    String line;
-    // We have seen that occasionally the Homo_sapiens_gene_info.gz
-    // contains duplicate lines, which is an error but we do not want the code
-    // to crash, so we check for previously found term ids with the seen set.
-    // The TermId <-> symbol mapping is one to one.
-    Set<TermId> seen = new HashSet<>();
-    while ((line=br.readLine())!=null) {
-      String[] a = line.split("\t");
-      String taxon=a[0];
-      if (! taxon.equals("9606")) continue; // i.e., we want only Homo sapiens sapiens and not Neaderthal etc.
-      if(!("unknown".equals(a[9]) | "tRNA".equals(a[9]) | "rRNA".equals(a[9]) | "pseudo".equals(a[9]))){
-        String geneId=a[1];
-        String symbol=a[2];
-        TermId tid = TermId.of(ENTREZ_GENE_PREFIX,geneId);
-        if (seen.contains(tid)) {
-          continue;
-        }
-        seen.add(tid);
-        builder.put(tid,symbol);
-      }
-    }
-    return allGeneIdToSymbolMap = builder.build();
-  }
+
 
 }
